Removed circular dependencies
Used std::weak_ptr where needed

+++ include/valuation/Program.cpp
@@ -3,7 +3,7 @@
 // Initialise static values
 unsigned long Program::programs = 0;
 
-Program::Program(const Eigen::VectorXd& flowRates, const Eigen::MatrixXf&
+Program::Program(const Eigen::VectorXd& flowRates, const Eigen::MatrixXd&
         switching) {
 
 	this->number = ++programs;
